requirements.txt
alembic==1.15.2
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
cloudinary==1.43.0
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
pillow==11.1.0
python-dotenv==1.1.0
python-slugify==8.0.4
six==1.17.0
SQLAlchemy==2.0.40
text-unidecode==1.3
typing_extensions==4.13.0
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
